# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pixelate",
        "description": "Documentation for Pixelate app",
        "version": "1.0.2"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Server"
        }
    ],
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "description": "Register new user",
                "operationId": "App\\Http\\Controllers\\Auth\\RegisterController::registered",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "username",
                                    "name",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Fanny"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Fanny Besoz"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "fanny256@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "pass1234"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "pass1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Fanny Anna"
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "Fanny"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "fanny256@email.com"
                                        },
                                        "id": {
                                            "type": "number",
                                            "example": 25
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2020-11-27T08:16:31.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2020-11-27T08:16:31.000000Z"
                                        },
                                        "photo_url": {
                                            "type": "string",
                                            "example": "https://www.gravatar.com/avatar/a9d770a5af9a113e77c887c0b772d98fjpg?s=200&d=mm"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "username": {
                                                    "example": [
                                                        "The username has already been taken."
                                                    ]
                                                },
                                                "email": {
                                                    "example": [
                                                        "The email has already been taken."
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Already Registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Please confirm your email"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "description": "Example of Auth Header Token (Copy and paste below to authorize)=>\nBearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sb2NhbGhvc3Q6ODAwMFwvYXBpXC9sb2dpbiIsImlhdCI6MTYwNTM0ODU1OSwiZXhwIjoxNjA1MzUyMTU5LCJuYmYiOjE2MDUzNDg1NTksImp0aSI6IjlqMDN0MVVISThoNjlhRUsiLCJzdWIiOjEsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.dn45EdZMatpW2zAQcZklDS2902mwVNx44TbkDiiGLXY",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}